+++ src/pyc_wheel/_pyc_wheel.py
@@ -58,7 +58,8 @@ def create_pyc_whl_path(source_whl: pathlib.Path) -> pathlib.Path:
     return pyc_whl
 
 
-def convert_wheel(whl_file: Path, *, exclude=None, with_backup=False, rename=False, quiet=False, optimize=0):
+def convert_wheel(whl_file: Path, *, exclude=None, with_backup=False, rename=False,
+                  quiet=False, optimize=0):
     """Generate a new whl with only pyc files."""
 
     if whl_file.suffix != ".whl":
@@ -113,12 +114,12 @@ def convert_wheel(whl_file: Path, *, exclude=None, with_backup=False, rename=Fal
             file_path = whl_path/member.filename
             timestamp = datetime(*member.date_time).timestamp()
             try:
-                os.utime(str(file_path), (timestamp, timestamp))
+                os.utime(file_path, (timestamp, timestamp))
             except Exception:
                 pass  # ignore errors
             permission_bits = (member.external_attr >> 16) & 0o777
             try:
-                os.chmod(str(file_path), permission_bits)
+                os.chmod(file_path, permission_bits)
             except Exception:
                 pass  # ignore errors
 
@@ -133,7 +134,7 @@ def convert_wheel(whl_file: Path, *, exclude=None, with_backup=False, rename=Fal
             whl_file.replace(whl_file.with_suffix(whl_file.suffix + ".bak"))
         if rename:
             pyc_whl_path = create_pyc_whl_path(whl_file)
-            whl_file_zip.replace(pyc_whl_path)
+            shutil.move(whl_file_zip, pyc_whl_path)
             if whl_file != pyc_whl_path:
                 whl_file.unlink(missing_ok=True)
                 if rename == "symlink":
@@ -142,7 +143,7 @@ def convert_wheel(whl_file: Path, *, exclude=None, with_backup=False, rename=Fal
                     print("Renamed wheel: {!s} -> {!s}".format(whl_file, pyc_whl_path))
             whl_file = pyc_whl_path
         else:
-            whl_file_zip.replace(whl_file)
+            shutil.move(whl_file_zip, whl_file)
         return whl_file
     finally:
         # Clean up original directory
@@ -211,16 +212,16 @@ def rewrite_dist_info(dist_info_path: Path, *, exclude=None):
             tag_components[0] = create_python_tag()
             pyc_tag = "-".join(tag_components)
             break
-        # if (python_tag == f"cp{py_major_version}{py_minor_version}"
-        #    and py_implementation == "CPython"):
-        #     tag_components[0] = create_python_tag()
-        #     pyc_tag = "-".join(tag_components)
-        #     break
-        # if (python_tag == f"pp{py_major_version}{py_minor_version}"
-        #    and py_implementation == "PyPy"):
-        #     tag_components[0] = create_python_tag()
-        #     pyc_tag = "-".join(tag_components)
-        #     break
+        if (python_tag == f"cp{py_major_version}{py_minor_version}"
+           and py_implementation == "CPython"):
+            tag_components[0] = create_python_tag()
+            pyc_tag = "-".join(tag_components)
+            break
+        if (python_tag == f"pp{py_major_version}{py_minor_version}"
+           and py_implementation == "PyPy"):
+            tag_components[0] = create_python_tag()
+            pyc_tag = "-".join(tag_components)
+            break
 
     if pyc_tag is None:
         raise RuntimeError("Cannot convert wheel with the used interpreter.")
@@ -263,8 +264,10 @@ def main(argv=sys.argv[1:]):
     rename_group.add_argument("--rename", default=False, action="store_true",
                               help="Rename the wheel to python version.")
     if hasattr(os, "symlink"):
-        rename_group.add_argument("--symlink", dest="rename", action="store_const", const="symlink",
-                                  help="Rename the wheel to python version and symlink old name to new.")
+        rename_group.add_argument("--symlink", dest="rename", action="store_const",
+                                  const="symlink",
+                                  help="Rename the wheel to python version and symlink "
+                                       "old name to new.")
     parser.add_argument("--optimize", default=0, type=int, choices=[0, 1, 2],
                         help="Specifies the optimization level of the compiler."
                              "Explicit levels are 0 (no optimization; __debug__ is true),"